requirements.txt
pillow
ipdb
ipython
numpy
pylama
pylama-pylint